Error handling (#59)
* feat: :goal_net: error handling concept

* feat: :goal_net: error handling

+++ packages/server/src/schema/employment/mutations.ts
@@ -1,12 +1,20 @@
+import * as grpc from "@grpc/grpc-js";
 import client from "@gymlabs/admin.grpc.client";
 import {
-  BooleanType,
+  BooleanType__Output,
   Employment__Output,
 } from "@gymlabs/admin.grpc.definition";
 import { ZodError } from "zod";
 
 import { Employment } from "./types";
 import { builder } from "../builder";
+import {
+  InternalServerError,
+  InvalidArgumentError,
+  NotFoundError,
+  UnauthenticatedError,
+  UnauthorizedError,
+} from "../errors";
 
 builder.mutationFields((t) => ({
   createEmployment: t.fieldWithInput({
@@ -16,25 +24,45 @@ builder.mutationFields((t) => ({
       userId: t.input.string(),
       role: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        UnauthenticatedError,
+        InvalidArgumentError,
+        InternalServerError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      // TODO: check access rights
-      const employment: Employment__Output = await new Promise(
-        (resolve, reject) => {
-          client.createEmployment(input, (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const employment: Employment__Output = await new Promise(
+          (resolve, reject) => {
+            client.createEmployment(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return {
+          ...employment,
+          createdAt: new Date(employment.createdAt),
+          updatedAt: new Date(employment.updatedAt),
+        };
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return {
-        ...employment,
-        createdAt: new Date(employment.createdAt),
-        updatedAt: new Date(employment.updatedAt),
-      };
+      }
     },
   }),
 
@@ -43,19 +71,44 @@ builder.mutationFields((t) => ({
     input: {
       employmentId: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InvalidArgumentError,
+        NotFoundError,
+        InternalServerError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      // TODO: check access rights
-      const success: BooleanType = await new Promise((resolve, reject) => {
-        client.activateEmployment(input, (err, res) => {
-          if (err || !res) {
-            reject(err);
-          } else {
-            resolve(res);
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const success: BooleanType__Output = await new Promise(
+          (resolve, reject) => {
+            client.activateEmployment(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
           }
-        });
-      });
-      return success.value ?? false;
+        );
+        return success.value;
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
+        }
+      }
     },
   }),
 
@@ -65,19 +118,44 @@ builder.mutationFields((t) => ({
       gymId: t.input.string(),
       userId: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InvalidArgumentError,
+        NotFoundError,
+        InternalServerError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      // TODO: check access rights
-      const success: BooleanType = await new Promise((resolve, reject) => {
-        client.deleteEmployment(input, (err, res) => {
-          if (err || !res) {
-            reject(err);
-          } else {
-            resolve(res);
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const success: BooleanType__Output = await new Promise(
+          (resolve, reject) => {
+            client.deleteEmployment(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
           }
-        });
-      });
-      return success.value ?? false;
+        );
+        return success.value;
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
+        }
+      }
     },
   }),
 }));
