+++ packages/server/src/schema/exercises/mutations.ts
@@ -1,3 +1,4 @@
+import * as grpc from "@grpc/grpc-js";
 import client from "@gymlabs/admin.grpc.client";
 import {
   BooleanType__Output,
@@ -8,6 +9,13 @@ import { ZodError } from "zod";
 
 import { mapNullToUndefined } from "../../lib/mapNullToUndefined";
 import { builder } from "../builder";
+import {
+  InternalServerError,
+  InvalidArgumentError,
+  NotFoundError,
+  UnauthenticatedError,
+  UnauthorizedError,
+} from "../errors";
 import { Exercise, ExerciseStep } from "../exercises/types";
 
 builder.mutationFields((t) => ({
@@ -18,29 +26,50 @@ builder.mutationFields((t) => ({
       name: t.input.string(),
       description: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InvalidArgumentError,
+        InternalServerError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      const exercise: Exercise__Output = await new Promise(
-        (resolve, reject) => {
-          client.createExercise(input, (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const exercise: Exercise__Output = await new Promise(
+          (resolve, reject) => {
+            client.createExercise(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return {
+          ...exercise,
+          steps: exercise.steps.map((step) => ({
+            ...step,
+            createdAt: new Date(step.createdAt),
+            updatedAt: new Date(step.updatedAt),
+          })),
+          createdAt: new Date(exercise.createdAt),
+          updatedAt: new Date(exercise.updatedAt),
+        };
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return {
-        ...exercise,
-        steps: exercise.steps.map((step) => ({
-          ...step,
-          createdAt: new Date(step.createdAt),
-          updatedAt: new Date(step.updatedAt),
-        })),
-        createdAt: new Date(exercise.createdAt),
-        updatedAt: new Date(exercise.updatedAt),
-      };
+      }
     },
   }),
   updateExercise: t.fieldWithInput({
@@ -50,29 +79,53 @@ builder.mutationFields((t) => ({
       name: t.input.string({ required: false }),
       description: t.input.string({ required: false }),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InternalServerError,
+        InvalidArgumentError,
+        NotFoundError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      const exercise: Exercise__Output = await new Promise(
-        (resolve, reject) => {
-          client.updateExercise(mapNullToUndefined(input), (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const exercise: Exercise__Output = await new Promise(
+          (resolve, reject) => {
+            client.updateExercise(mapNullToUndefined(input), (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return {
+          ...exercise,
+          steps: exercise.steps.map((step) => ({
+            ...step,
+            createdAt: new Date(step.createdAt),
+            updatedAt: new Date(step.updatedAt),
+          })),
+          createdAt: new Date(exercise.createdAt),
+          updatedAt: new Date(exercise.updatedAt),
+        };
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return {
-        ...exercise,
-        steps: exercise.steps.map((step) => ({
-          ...step,
-          createdAt: new Date(step.createdAt),
-          updatedAt: new Date(step.updatedAt),
-        })),
-        createdAt: new Date(exercise.createdAt),
-        updatedAt: new Date(exercise.updatedAt),
-      };
+      }
     },
   }),
   deleteExercise: t.fieldWithInput({
@@ -80,20 +133,44 @@ builder.mutationFields((t) => ({
     input: {
       id: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InvalidArgumentError,
+        NotFoundError,
+        InternalServerError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      const exercise: BooleanType__Output = await new Promise(
-        (resolve, reject) => {
-          client.deleteExercise(input, (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const exercise: BooleanType__Output = await new Promise(
+          (resolve, reject) => {
+            client.deleteExercise(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return exercise.value;
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return exercise.value;
+      }
     },
   }),
   createExerciseStep: t.fieldWithInput({
@@ -104,24 +181,45 @@ builder.mutationFields((t) => ({
       name: t.input.string(),
       description: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InternalServerError,
+        InvalidArgumentError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      const exerciseStep: ExerciseStep__Output = await new Promise(
-        (resolve, reject) => {
-          client.createExerciseStep(input, (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const exerciseStep: ExerciseStep__Output = await new Promise(
+          (resolve, reject) => {
+            client.createExerciseStep(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return {
+          ...exerciseStep,
+          createdAt: new Date(exerciseStep.createdAt),
+          updatedAt: new Date(exerciseStep.updatedAt),
+        };
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return {
-        ...exerciseStep,
-        createdAt: new Date(exerciseStep.createdAt),
-        updatedAt: new Date(exerciseStep.updatedAt),
-      };
+      }
     },
   }),
   updateExerciseStep: t.fieldWithInput({
@@ -132,24 +230,48 @@ builder.mutationFields((t) => ({
       description: t.input.string({ required: false }),
       index: t.input.int({ required: false }),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InternalServerError,
+        InvalidArgumentError,
+        NotFoundError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      const exerciseStep: ExerciseStep__Output = await new Promise(
-        (resolve, reject) => {
-          client.updateExerciseStep(mapNullToUndefined(input), (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const exerciseStep: ExerciseStep__Output = await new Promise(
+          (resolve, reject) => {
+            client.updateExerciseStep(mapNullToUndefined(input), (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return {
+          ...exerciseStep,
+          createdAt: new Date(exerciseStep.createdAt),
+          updatedAt: new Date(exerciseStep.updatedAt),
+        };
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return {
-        ...exerciseStep,
-        createdAt: new Date(exerciseStep.createdAt),
-        updatedAt: new Date(exerciseStep.updatedAt),
-      };
+      }
     },
   }),
   deleteExerciseStep: t.fieldWithInput({
@@ -157,20 +279,44 @@ builder.mutationFields((t) => ({
     input: {
       id: t.input.string(),
     },
-    errors: { types: [ZodError] },
+    errors: {
+      types: [
+        ZodError,
+        InvalidArgumentError,
+        InternalServerError,
+        NotFoundError,
+        UnauthenticatedError,
+        UnauthorizedError,
+      ],
+    },
     resolve: async (query, { input }, args, context) => {
-      const exerciseStep: BooleanType__Output = await new Promise(
-        (resolve, reject) => {
-          client.deleteExerciseStep(input, (err, res) => {
-            if (err || !res) {
-              reject(err);
-            } else {
-              resolve(res);
-            }
-          });
+      if (!args.viewer.isAuthenticated()) throw new UnauthenticatedError();
+      try {
+        const exerciseStep: BooleanType__Output = await new Promise(
+          (resolve, reject) => {
+            client.deleteExerciseStep(input, (err, res) => {
+              if (err) {
+                reject(err);
+              } else if (res) {
+                resolve(res);
+              }
+            });
+          }
+        );
+        return exerciseStep.value;
+      } catch (err) {
+        const error = err as grpc.ServiceError;
+        switch (error.code) {
+          case grpc.status.INVALID_ARGUMENT:
+            throw new InvalidArgumentError(error.message);
+          case grpc.status.NOT_FOUND:
+            throw new NotFoundError(error.message);
+          case grpc.status.PERMISSION_DENIED:
+            throw new UnauthorizedError();
+          default:
+            throw new InternalServerError();
         }
-      );
-      return exerciseStep.value;
+      }
     },
   }),
 }));
